refactor: rename InitWizard files
The main class has been renamed from InitWizard to InitWizard::Wizard.

I might change it again in the future, but it's going to be that for
now.
